Redact token when printed on screen
When the token is printed it's generally to let the user know that it has *some* value, rather than needing to expose the exact value. We now display the token in redacted form when it's shown on-screen, with the exception of `doppler configure get token`, `doppler configure debug`, and commands run with the `--json` flag.