Add support for mounting secrets to file path
When mounted, secrets are written to an ephemeral file instead of being injected into the environment. This file is available for the duration of the 'run' command.

Technically, the mounted secrets file is actually a named pipe. This allows us to ensure that the pipe's contents (i.e. the secrets) are inaccessible once the CLI exits. We still attempt to clean up this file on exit, but that's not guaranteed to  run due to the ability to SIGKILL the CLI.