chore: fix platform error when building on arm64 machine
When our docker image is built on an arm64 machine, it fails to execute on amd64 machines. Conversely, when our image is built on an amd64 machine, it can run on both architecures. Thus, we now always build an amd64 image. This is an intermediate step until we ultimately build an image for each architecture.