Use standard zsh path for rpm/RHEL releases
Due to a quirk with OSTree releases, the rpm releases should place the doppler zsh completion file in /usr/share/zsh/, which is the default for Fedora/RHEL/CentOS based systems. In theory this should fix #331 but I will need to perform testing to confirm. There may be a more elegant way to accomplish this, but this seems to be the most straight forward way that I can find for the moment.