chore: replace slices.Contains with utils.Contains
We were getting the following gosec error when using `slices.Contains`:

```
could not import slices (invalid package name: "")
```

This is because:

- `slices` was added in Go 1.21
- Support for Go 1.21 was added in gosec 2.17
- We are currently using gosec 2.15.0 and via salus 3.2.5 via federacy/scan-action which pulls `coinbase/salus:latest`

Salus 3.2.6 has been released with the gosec version bump but it isn't tagged `latest`. I've filed https://github.com/coinbase/salus/issues/880 to address.

For now, the easiest thing to do was just to not use the `slices` module, which was easy because we already have a util which does the same thing.